+++ lib/espnscrape.rb
@@ -5,8 +5,15 @@ require 'nokogiri'
 require 'open-uri'
 require 'time'
 
+# Core Extensions
+require_relative 'espnscrape/Hash'
+require_relative 'espnscrape/Struct'
+require_relative 'espnscrape/String'
+
 # To Hash | Struct conversions
 require_relative 'espnscrape/ArrayConversions'
+
+require_relative 'espnscrape/Navigator'
 
 # espnscrape
 require_relative 'espnscrape/NbaUrls'
@@ -25,15 +32,20 @@ include PrintUtils
 # EspnScrape main class
 class EspnScrape
   # Gem Version
-  VERSION = '0.4.0'.freeze
+  VERSION = '0.5.0'.freeze
+  # initialize
+  def initialize(config = {})
+    @format = defaultFormat(config[:format])
+  end
 
   # Returns an {NbaBoxScore} object
   # @param game_id [Integer] Boxscore ID
   # @return [NbaBoxScore] NbaBoxScore
   # @example
   #   EspnScrape.boxscore(493848273)
-  def self.boxscore(game_id)
-    NbaBoxScore.new game_id
+  def self.boxscore(game_id, f_mat = nil)
+    NbaBoxScore.new(game_id: game_id,
+                    format: defaultFormat(f_mat))
   end
 
   # Returns an {NbaBoxScore} object
@@ -41,8 +53,8 @@ class EspnScrape
   # @return (see .boxscore)
   # @example
   #  es.boxscore(493848273)
-  def boxscore(game_id)
-    EspnScrape.boxscore game_id
+  def boxscore(game_id, f_mat = nil)
+    EspnScrape.boxscore game_id, (f_mat || @format)
   end
 
   # Returns an {NbaRoster} object
@@ -50,8 +62,9 @@ class EspnScrape
   # @return [NbaRoster] NbaRoster
   # @example
   #   EspnScrape.roster('UTA')
-  def self.roster(team_id)
-    NbaRoster.new team_id
+  def self.roster(team_id, f_mat = nil)
+    NbaRoster.new(team_id: team_id,
+                  format: defaultFormat(f_mat))
   end
 
   # Returns an {NbaRoster} object
@@ -59,24 +72,24 @@ class EspnScrape
   # @return (see .roster)
   # @example
   #  es.roster('UTA')
-  def roster(team_id)
-    EspnScrape.roster team_id
+  def roster(team_id, f_mat = nil)
+    EspnScrape.roster team_id, (f_mat || @format)
   end
 
   # Return Array of Team Data
   # @return [[[String]]] NBA Team Data
   # @example
   #  EspnScrape.teamList
-  def self.teamList
-    NbaTeamList.new.teamList
+  def self.teamList(f_mat = nil)
+    NbaTeamList.new(format: defaultFormat(f_mat)).teamList
   end
 
   # Return Array of Team Data
   # @return (see .teamList)
   # @example
   #  es.teamList
-  def teamList
-    EspnScrape.teamList
+  def teamList(f_mat = nil)
+    EspnScrape.teamList(f_mat || @format)
   end
 
   # Return an {NbaSchedule} object
@@ -86,8 +99,10 @@ class EspnScrape
   # @example
   #   EspnScrape.schedule('UTA')    # Schedule for Latest Season Type
   #   EspnScrape.schedule('TOR', 3) # Playoff Schedule
-  def self.schedule(team_id, s_type = '')
-    NbaSchedule.new team_id, '', s_type
+  def self.schedule(team_id, s_type = '', f_mat = nil)
+    NbaSchedule.new(team_id: team_id,
+                    season_type: s_type,
+                    format: defaultFormat(f_mat))
   end
 
   # Return an {NbaSchedule} object
@@ -96,8 +111,8 @@ class EspnScrape
   # @example
   #  es.schedule('MIA')     # Schedule for Latest Season Type
   #  es.schedule('DET', 1)  # Preseason Schedule
-  def schedule(team_id, s_type = '')
-    EspnScrape.schedule team_id, s_type
+  def schedule(team_id, s_type = '', f_mat = nil)
+    EspnScrape.schedule team_id, s_type, (f_mat || @format)
   end
 
   # Return new {NbaPlayer} object